Conflicts:
	Modules/FindCUDA.cmake
	Modules/FindCUDA/make2cmake.cmake
	Modules/FindCUDA/parse_cubin.cmake
	Modules/FindCUDA/run_nvcc.cmake

+++ Tests/RunCMake/RunCMake.cmake
@@ -26,12 +26,22 @@ function(run_cmake test)
     endif()
   endforeach()
   set(RunCMake_TEST_SOURCE_DIR "${top_src}")
-  set(RunCMake_TEST_BINARY_DIR "${top_bin}/${test}-build")
-  file(REMOVE_RECURSE "${RunCMake_TEST_BINARY_DIR}")
+  if(NOT RunCMake_TEST_BINARY_DIR)
+    set(RunCMake_TEST_BINARY_DIR "${top_bin}/${test}-build")
+  endif()
+  if(NOT RunCMake_TEST_NO_CLEAN)
+    file(REMOVE_RECURSE "${RunCMake_TEST_BINARY_DIR}")
+  endif()
   file(MAKE_DIRECTORY "${RunCMake_TEST_BINARY_DIR}")
+  if(NOT DEFINED RunCMake_TEST_OPTIONS)
+    set(RunCMake_TEST_OPTIONS "")
+  endif()
   execute_process(
     COMMAND ${CMAKE_COMMAND} "${RunCMake_TEST_SOURCE_DIR}"
-              -G "${RunCMake_GENERATOR}" -DRunCMake_TEST=${test}
+              -G "${RunCMake_GENERATOR}"
+              -T "${RunCMake_GENERATOR_TOOLSET}"
+              -DRunCMake_TEST=${test}
+              ${RunCMake_TEST_OPTIONS}
     WORKING_DIRECTORY "${RunCMake_TEST_BINARY_DIR}"
     OUTPUT_VARIABLE actual_stdout
     ERROR_VARIABLE actual_stderr
